Move mda module mda_sizes into sizes module
