+++ src/engine/strat_engine/backstore/metadata/mod.rs
@@ -4,5 +4,6 @@
 
 mod bda;
 mod mda;
+mod sizes;
 
-pub use self::{bda::BDA, mda::MDADataSize};
+pub use self::{bda::BDA, sizes::MDADataSize};
